h264: silence remaining clippy warnings so -D warnings passes clean
Cleans up ~23 residual lints after the earlier cargo clippy --fix pass,
plus one corrupted doc comment left behind by that run.

Production code fixes (no behaviour change):

- src/cavlc_enc.rs: reindent doc list continuation.
- src/deblock.rs: collapse the identical 4:4:4 / 4:2:2 chroma-edge-row
  branches into a single arm (doc-cites both paths).
- src/dpb.rs: sort_by(|a, b| b.key.cmp(&a.key)) -> sort_by_key(|b|
  Reverse(b.key)) in the list0_b / list0_p / list1_b helpers.
- src/intra_pred.rs + src/intra_pred_hi.rs: convert the ftl
  late-init / per-branch assignment into a single let binding in
  filter_ref_8x8 (both 8- and 10-bit variants).
- src/motion.rs: unnecessary_unwrap via if-let destructuring for the
  "B and C unavailable, A available" special cases (predict_mv_l0 /
  predict_mv_list); match-block -> matches!() in P_Skip trivial-A/B
  detection; rewrap the luma_mc_plane doc so the leading "+ bilinear"
  line doesn't get parsed as a markdown list continuation.
- src/b_mb.rs: repair mangled `> > §8.4.1.2.3` doc block from the
  earlier auto-fix pass back to a single spec pointer.

Test-only fixes:

- src/intra_pred.rs tests module: allow identity_op + erasing_op. The
  test bodies use `out[y * stride + x]` with `y` / `x` pinned at spec
  indices like (0, 1) or (3, 0). Collapsing "0 * 8 + 1" to "1" would
  hide which zHU / zVR the assert targets, so the spec-formula form
  stays.
- tests/{pb_residual_regression,yuv444_ext}.rs: collect_frames loops
  converted to while-let-Ok.
- tests/decode_bframe{,_temporal,_implicit}.rs: needless_range_loop ->
  enumerate().take(cmp_count).
- tests/decode_b_10bit.rs: same fix, and switch the inner write to
  the mutable iter slot so the elided index still ends up assigned.
- tests/encode_pframe_roundtrip.rs: manual clamp -> i32::clamp.

cargo clippy --release --all-targets -- -D warnings exits 0.
cargo test --release: 224 passed, 0 failed, 2 ignored (unchanged from
pre-fix baseline; no regressions).

+++ src/transform.rs
@@ -85,7 +85,7 @@ pub fn dequantize_4x4_scaled(coeffs: &mut [i32; 16], qp: i32, weight_scale: &[i1
         }
         return;
     }
-    let qp6 = (qp / 6) as i32;
+    let qp6 = qp / 6;
     let qmod = (qp % 6) as usize;
     if qp6 >= 4 {
         let shift = (qp6 - 4) as u32;
@@ -166,7 +166,7 @@ pub fn idct_4x4(coeffs: &mut [i32; 16]) {
 /// §8.5.12.1 formula with an unbounded `qP/6` shift.
 pub fn dequantize_4x4_scaled_ext(coeffs: &mut [i32; 16], qp: i32, weight_scale: &[i16; 16]) {
     let qp = qp.max(0);
-    let qp6 = (qp / 6) as i32;
+    let qp6 = qp / 6;
     let qmod = (qp % 6) as usize;
     if weight_scale == &crate::scaling_list::FLAT_4X4 {
         let shift = qp6 as u32;
@@ -230,7 +230,7 @@ pub fn inv_hadamard_4x4_dc_scaled_ext(dc: &mut [i32; 16], qp: i32, weight_scale_
     }
 
     let qp = qp.max(0);
-    let qp6 = (qp / 6) as i32;
+    let qp6 = qp / 6;
     let qmod = (qp % 6) as usize;
     let v = V_TABLE[qmod][0] as i64;
     if weight_scale_00 == 16 {
@@ -296,12 +296,12 @@ pub fn inv_hadamard_2x2_chroma_dc_scaled_ext(dc: &mut [i32; 4], qp: i32, weight_
     } else {
         let w = weight_scale_00 as i64;
         if qp >= 30 {
-            let shift = (qp6 - 5) as u32;
+            let shift = qp6 - 5;
             for x in dc.iter_mut() {
                 *x = ((*x as i64 * w * v) << shift) as i32;
             }
         } else {
-            let shift = (5 - qp6) as u32;
+            let shift = 5 - qp6;
             let round = 1i64 << (shift - 1);
             for x in dc.iter_mut() {
                 *x = ((*x as i64 * w * v + round) >> shift) as i32;
@@ -373,7 +373,7 @@ pub fn inv_hadamard_4x4_dc_scaled(dc: &mut [i32; 16], qp: i32, weight_scale_00: 
     // `2 - qP/6` right shift with rounding). The scaled path below
     // generalises by factoring `w` into the intermediate.
     let qp = qp.clamp(0, 51);
-    let qp6 = (qp / 6) as i32;
+    let qp6 = qp / 6;
     let qmod = (qp % 6) as usize;
     let v = V_TABLE[qmod][0];
     if weight_scale_00 == 16 {
@@ -471,7 +471,7 @@ pub fn inv_hadamard_2x4_chroma_dc_scaled(dc: &mut [i32; 8], qp: i32, weight_scal
     // same shape as the 2×2 code path, with QP replaced by QP + 3 and one
     // more right-shift absorbed into the shift constants.
     let qp_dc = (qp + 3).clamp(0, 51);
-    let qp6 = (qp_dc / 6) as i32;
+    let qp6 = qp_dc / 6;
     let qmod = (qp_dc % 6) as usize;
     let v = V_TABLE[qmod][0];
     let w = weight_scale_00 as i32;
@@ -513,7 +513,7 @@ pub fn inv_hadamard_2x4_chroma_dc_scaled_ext(dc: &mut [i32; 8], qp: i32, weight_
     }
 
     let qp_dc = (qp + 3).max(0);
-    let qp6 = (qp_dc / 6) as i32;
+    let qp6 = qp_dc / 6;
     let qmod = (qp_dc % 6) as usize;
     let v = V_TABLE[qmod][0] as i64;
     let w = weight_scale_00 as i64;
@@ -575,12 +575,12 @@ pub fn inv_hadamard_2x2_chroma_dc_scaled(dc: &mut [i32; 4], qp: i32, weight_scal
         // §8.5.11.1 full form: (f * w * V + rnd) >> 5 for qP < 30,
         // shift-left for qP >= 30.
         if qp >= 30 {
-            let shift = (qp6 - 5) as u32;
+            let shift = qp6 - 5;
             for x in dc.iter_mut() {
                 *x = (*x * w * v) << shift;
             }
         } else {
-            let shift = (5 - qp6) as u32;
+            let shift = 5 - qp6;
             let round = 1i32 << (shift - 1);
             for x in dc.iter_mut() {
                 *x = (*x * w * v + round) >> shift;
@@ -685,7 +685,7 @@ pub fn dequantize_8x8(coeffs: &mut [i32; 64], qp: i32) {
 /// `w`.
 pub fn dequantize_8x8_scaled(coeffs: &mut [i32; 64], qp: i32, weight_scale: &[i16; 64]) {
     let qp = qp.clamp(0, 51);
-    let qp6 = (qp / 6) as i32;
+    let qp6 = qp / 6;
     let qmod = (qp % 6) as usize;
     if qp6 >= 6 {
         let shift = (qp6 - 6) as u32;
@@ -716,7 +716,7 @@ pub fn dequantize_8x8_scaled(coeffs: &mut [i32; 64], qp: i32, weight_scale: &[i1
 /// 10-bit P/B inter 8×8 residual paths.
 pub fn dequantize_8x8_scaled_ext(coeffs: &mut [i32; 64], qp: i32, weight_scale: &[i16; 64]) {
     let qp = qp.max(0);
-    let qp6 = (qp / 6) as i32;
+    let qp6 = qp / 6;
     let qmod = (qp % 6) as usize;
     if qp6 >= 6 {
         let shift = (qp6 - 6) as u32;
